+++ Wizard Soup Party/Assets/Art/Materials/grassyTexture.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 50df82c35065c1041a41cf18c87add3b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
