+++ Wizard Soup Party/Assets/Prefabs/spriteBillboard.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fae78270630cde44da25be105c868e81
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
