+++ Wizard Soup Party/Assets/Scenes/Test - playground.unity
@@ -161,6 +161,9 @@ MonoBehaviour:
   isGliding: 0
   glideGravity: -2
   maxGlideTime: 2
+  dashSpeed: 6
+  dashDuration: 0.05
+  dashCooldown: 1
 --- !u!143 &278324490
 CharacterController:
   m_ObjectHideFlags: 0
@@ -846,7 +849,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: 31321ba15b8f8eb4c954353edc038b1d, type: 2}
+  - {fileID: 2100000, guid: 50df82c35065c1041a41cf18c87add3b, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -886,7 +889,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalScale: {x: 4, y: 1, z: 4}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
