--- Wizard Soup Party/Assets/Scripts/Billboard.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4972d9864ffd42c45b12fbe8b16f6517
