+++ Wizard Soup Party/Assets/Scripts/Sprite Billboard Script.cs
@@ -1,0 +1,34 @@
+using System;
+using Unity.Cinemachine;
+using UnityEngine;
+
+
+public class SpriteBillboardScript : MonoBehaviour
+{
+
+    public CinemachineCamera cam;
+    private SpriteRenderer SR;
+    
+    private void Start()
+    {
+        SR = GetComponent<SpriteRenderer>();
+    }
+
+
+    // Update is called once per frame
+    void Update()
+    {
+        Vector3 targetPosition =  cam.transform.position;
+        
+        Vector3 direction = (transform.position - targetPosition).normalized;
+        direction.y = 0;
+
+        if (direction.magnitude > 0.1f)
+        {
+            transform.forward = direction.normalized;
+        }
+        
+        
+        //transform.rotation = Quaternion.Euler(0f, cam.transform.rotation.eulerAngles.y, 0f);
+    }
+}
