+++ Wizard Soup Party/Assets/Scripts/Sprite Billboard Script.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 839817d0b3083e245959730c85b05ae8
